requirements.txt
environs==9.5.0
Pillow==8.0.0
